refactor: CookieManager 유틸 클래스로 cookie 관련 메서드 관리 (#183)
